Lame license

+++ commons/src/main/kotlin/com/simplemobiletools/commons/models/License.kt
@@ -1,3 +1,3 @@
 package com.simplemobiletools.commons.models
 
-data class License(val id: Int, val titleId: Int, val textId: Int, val urlId: Int)
+data class License(val id: Long, val titleId: Int, val textId: Int, val urlId: Int)
